+++ src/components/Layout/index.tsx
@@ -1,0 +1,3 @@
+import Layout from "./Layout"
+
+export default Layout
